TACT-347: Refactored folders for pages
