+++ pages/goals/index.tsx
@@ -1,7 +1,7 @@
 import dynamic from 'next/dynamic';
 import { PageLoader } from '../../components/shared/PageLoader';
 
-const DynamicGoalsPage = dynamic(() => import('../../components/pages/Goals'), {
+const DynamicGoalsPage = dynamic(() => import('../../components/pages/Goals/MainPage'), {
   loading: () => <PageLoader />,
   ssr: false,
 });
